agents/projects/my_assistant/requirements.txt
openai
chromadb
gradio
python-dotenv
pydantic
requests
pypdf